pyproject.toml
[project]
name = "generate-recipe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.4",
    "streamlit>=1.43.2",
]
